REVIEW: externalize tests
