+++ versioned/persist/tests/src/main/java/org/projectnessie/versioned/persist/tests/AbstractConcurrency.java
@@ -1,0 +1,235 @@
+/*
+ * Copyright (C) 2020 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.versioned.persist.tests;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.google.protobuf.ByteString;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.projectnessie.versioned.BranchName;
+import org.projectnessie.versioned.Key;
+import org.projectnessie.versioned.ReferenceRetryFailureException;
+import org.projectnessie.versioned.persist.adapter.ContentsAndState;
+import org.projectnessie.versioned.persist.adapter.ContentsId;
+import org.projectnessie.versioned.persist.adapter.DatabaseAdapter;
+import org.projectnessie.versioned.persist.adapter.ImmutableCommitAttempt;
+import org.projectnessie.versioned.persist.adapter.KeyFilterPredicate;
+import org.projectnessie.versioned.persist.adapter.KeyWithBytes;
+
+public abstract class AbstractConcurrency {
+
+  private final DatabaseAdapter databaseAdapter;
+
+  protected AbstractConcurrency(DatabaseAdapter databaseAdapter) {
+    this.databaseAdapter = databaseAdapter;
+  }
+
+  static class Variation {
+    final int threads = Math.max(4, Runtime.getRuntime().availableProcessors());
+    final boolean singleBranch;
+    final boolean sharedKeys;
+    final int tables;
+
+    Variation(boolean singleBranch, boolean sharedKeys, int tables) {
+      this.singleBranch = singleBranch;
+      this.sharedKeys = sharedKeys;
+      this.tables = tables;
+    }
+
+    @Override
+    public String toString() {
+      return "threads="
+          + threads
+          + ", singleBranch="
+          + singleBranch
+          + ", sharedKeys="
+          + sharedKeys
+          + ", tables="
+          + tables;
+    }
+  }
+
+  /** Cartesian product of all {@link Variation}s for {@link #concurrency(Variation)}. */
+  @SuppressWarnings("unused")
+  static Stream<Variation> concurrencyVariations() {
+    return Stream.of(Boolean.FALSE, Boolean.TRUE)
+        .flatMap(
+            singleBranch ->
+                Stream.of(Boolean.FALSE, Boolean.TRUE)
+                    .flatMap(
+                        sharedKeys ->
+                            Stream.of(3)
+                                .map(tables -> new Variation(singleBranch, sharedKeys, tables))));
+  }
+
+  @ParameterizedTest
+  @MethodSource("concurrencyVariations")
+  void concurrency(Variation variation) throws Exception {
+    ExecutorService executor = Executors.newFixedThreadPool(variation.threads);
+    AtomicInteger commitsOK = new AtomicInteger();
+    AtomicInteger retryFailures = new AtomicInteger();
+    AtomicBoolean stopFlag = new AtomicBoolean();
+    List<Runnable> tasks = new ArrayList<>(variation.threads);
+    Map<Key, ContentsId> keyToContentsId = new HashMap<>();
+    try {
+      CountDownLatch startLatch = new CountDownLatch(1);
+      CountDownLatch stopLatch = new CountDownLatch(variation.threads);
+      Map<BranchName, Set<Key>> keysPerBranch = new HashMap<>();
+      for (int i = 0; i < variation.threads; i++) {
+        BranchName branch =
+            BranchName.of("concurrency-" + ((variation.singleBranch ? "shared" : i)));
+        List<Key> keys = new ArrayList<>(variation.tables);
+
+        for (int k = 0; k < variation.tables; k++) {
+          Key key =
+              Key.of(
+                  "some",
+                  "key",
+                  variation.sharedKeys ? "shared" : Integer.toString(i),
+                  "table-" + k);
+          keys.add(key);
+          keyToContentsId.put(
+              key,
+              ContentsId.of(
+                  String.format(
+                      "%s-table-%d", variation.sharedKeys ? "shared" : Integer.toString(i), k)));
+          keysPerBranch.computeIfAbsent(branch, x -> new HashSet<>()).add(key);
+        }
+
+        tasks.add(
+            () -> {
+              try {
+                assertThat(startLatch.await(2, TimeUnit.SECONDS)).isTrue();
+
+                for (int commit = 0; ; commit++) {
+                  if (stopFlag.get()) {
+                    stopLatch.countDown();
+                    break;
+                  }
+
+                  List<ByteString> currentStates;
+                  try (Stream<ByteString> str =
+                      databaseAdapter
+                          .values(
+                              databaseAdapter.toHash(branch), keys, KeyFilterPredicate.ALLOW_ALL)
+                          .collect(Collectors.toList())
+                          .stream()
+                          .map(Optional::get)
+                          .map(ContentsAndState::getGlobalState)) {
+                    currentStates = str.collect(Collectors.toList());
+                  }
+
+                  ImmutableCommitAttempt.Builder commitAttempt = ImmutableCommitAttempt.builder();
+
+                  for (int ki = 0; ki < keys.size(); ki++) {
+                    Key key = keys.get(ki);
+                    ContentsId contentsId = keyToContentsId.get(key);
+                    commitAttempt.putGlobal(
+                        contentsId,
+                        ByteString.copyFromUtf8(
+                            Integer.toString(
+                                Integer.parseInt(currentStates.get(ki).toStringUtf8()) + 1)));
+                    if (!variation.sharedKeys) {
+                      commitAttempt.putExpectedStates(
+                          contentsId, Optional.of(currentStates.get(ki)));
+                    }
+                    commitAttempt.addPuts(
+                        KeyWithBytes.of(keys.get(ki), contentsId, (byte) 0, ByteString.EMPTY));
+                  }
+
+                  try {
+                    commitAttempt
+                        .commitToBranch(branch)
+                        .commitMetaSerialized(
+                            ByteString.copyFromUtf8(
+                                "commit #"
+                                    + commit
+                                    + " to "
+                                    + branch.getName()
+                                    + " something "
+                                    + ThreadLocalRandom.current().nextLong()));
+                    databaseAdapter.commit(commitAttempt.build());
+                    commitsOK.incrementAndGet();
+                  } catch (ReferenceRetryFailureException retry) {
+                    retryFailures.incrementAndGet();
+                  }
+                }
+              } catch (Exception e) {
+                e.printStackTrace();
+                throw new RuntimeException(e);
+              }
+            });
+      }
+
+      for (Entry<BranchName, Set<Key>> branchKeys : keysPerBranch.entrySet()) {
+        BranchName branch = branchKeys.getKey();
+        databaseAdapter.create(branch, databaseAdapter.toHash(BranchName.of("main")));
+        ImmutableCommitAttempt.Builder commitAttempt =
+            ImmutableCommitAttempt.builder()
+                .commitToBranch(branchKeys.getKey())
+                .commitMetaSerialized(
+                    ByteString.copyFromUtf8("initial commit for " + branch.getName()));
+        for (Key k : branchKeys.getValue()) {
+          ContentsId contentsId = keyToContentsId.get(k);
+          commitAttempt.addPuts(KeyWithBytes.of(k, contentsId, (byte) 0, ByteString.EMPTY));
+          commitAttempt.putGlobal(contentsId, ByteString.copyFromUtf8("0"));
+        }
+        databaseAdapter.commit(commitAttempt.build());
+      }
+
+      tasks.forEach(executor::submit);
+
+      startLatch.countDown();
+      Thread.sleep(1_500);
+      stopFlag.set(true);
+
+      // 30 seconds is long, but necessary to let transactional databases detect deadlocks, which
+      // cause Nessie-commit-retries.
+      assertThat(stopLatch.await(30, TimeUnit.SECONDS)).isTrue();
+
+    } finally {
+      stopFlag.set(true);
+
+      System.out.printf(
+          "AbstractTieredCommitsTest.concurrency - %s : Commits OK: %s  Retry-Failures: %s%n",
+          variation, commitsOK, retryFailures);
+
+      executor.shutdownNow();
+
+      // 30 seconds is long, but necessary to let transactional databases detect deadlocks, which
+      // cause Nessie-commit-retries.
+      assertThat(executor.awaitTermination(30, TimeUnit.SECONDS)).isTrue();
+    }
+  }
+}
