+++ versioned/persist/tests/src/main/java/org/projectnessie/versioned/persist/tests/AbstractMergeTransplant.java
@@ -1,0 +1,155 @@
+/*
+ * Copyright (C) 2020 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.versioned.persist.tests;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import com.google.protobuf.ByteString;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Optional;
+import java.util.stream.Stream;
+import org.junit.jupiter.api.Test;
+import org.projectnessie.versioned.BranchName;
+import org.projectnessie.versioned.Hash;
+import org.projectnessie.versioned.Key;
+import org.projectnessie.versioned.ReferenceConflictException;
+import org.projectnessie.versioned.persist.adapter.CommitLogEntry;
+import org.projectnessie.versioned.persist.adapter.ContentsId;
+import org.projectnessie.versioned.persist.adapter.DatabaseAdapter;
+import org.projectnessie.versioned.persist.adapter.ImmutableCommitAttempt;
+import org.projectnessie.versioned.persist.adapter.KeyWithBytes;
+
+public abstract class AbstractMergeTransplant {
+
+  private final DatabaseAdapter databaseAdapter;
+
+  protected AbstractMergeTransplant(DatabaseAdapter databaseAdapter) {
+    this.databaseAdapter = databaseAdapter;
+  }
+
+  @Test
+  void merge() throws Exception {
+    mergeTransplant(
+        (target, expectedHead, branch, commitHashes, i) ->
+            databaseAdapter.merge(commitHashes[i], target, expectedHead));
+
+    BranchName branch = BranchName.of("branch");
+    BranchName branch2 = BranchName.of("branch2");
+    databaseAdapter.create(branch2, databaseAdapter.toHash(branch));
+    assertThatThrownBy(
+            () -> databaseAdapter.merge(databaseAdapter.toHash(branch), branch2, Optional.empty()))
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessageStartingWith("No hashes to merge from '");
+  }
+
+  @Test
+  void transplant() throws Exception {
+    Hash[] commits =
+        mergeTransplant(
+            (target, expectedHead, branch, commitHashes, i) ->
+                databaseAdapter.transplant(
+                    target, expectedHead, Arrays.asList(commitHashes).subList(0, i + 1)));
+
+    BranchName conflict = BranchName.of("conflict");
+
+    // no conflict, when transplanting the commits from against the current HEAD of the
+    // conflict-branch
+    Hash noConflictHead = databaseAdapter.toHash(conflict);
+    databaseAdapter.transplant(conflict, Optional.of(noConflictHead), Arrays.asList(commits));
+
+    // again, no conflict (same as above, just again)
+    databaseAdapter.transplant(conflict, Optional.empty(), Arrays.asList(commits));
+
+    assertThatThrownBy(
+            () -> databaseAdapter.transplant(conflict, Optional.empty(), Collections.emptyList()))
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessage("No hashes to transplant given.");
+  }
+
+  @FunctionalInterface
+  interface MergeOrTransplant {
+    void apply(
+        BranchName target,
+        Optional<Hash> expectedHead,
+        BranchName branch,
+        Hash[] commitHashes,
+        int i)
+        throws Exception;
+  }
+
+  private Hash[] mergeTransplant(MergeOrTransplant mergeOrTransplant) throws Exception {
+    BranchName main = BranchName.of("main");
+    BranchName branch = BranchName.of("branch");
+    BranchName conflict = BranchName.of("conflict");
+
+    databaseAdapter.create(branch, databaseAdapter.toHash(main));
+
+    Hash[] commits = new Hash[3];
+    for (int i = 0; i < commits.length; i++) {
+      ImmutableCommitAttempt.Builder commit =
+          ImmutableCommitAttempt.builder()
+              .commitToBranch(branch)
+              .commitMetaSerialized(ByteString.copyFromUtf8("commit " + i));
+      for (int k = 0; k < 3; k++) {
+        commit.addPuts(
+            KeyWithBytes.of(
+                Key.of("key", Integer.toString(k)),
+                ContentsId.of("C" + k),
+                (byte) 0,
+                ByteString.copyFromUtf8("value " + i + " for " + k)));
+      }
+      commits[i] = databaseAdapter.commit(commit.build());
+    }
+
+    for (int i = 0; i < commits.length; i++) {
+      BranchName target = BranchName.of("transplant-" + i);
+      databaseAdapter.create(target, databaseAdapter.toHash(main));
+
+      mergeOrTransplant.apply(target, Optional.empty(), branch, commits, i);
+
+      try (Stream<CommitLogEntry> targetLog =
+          databaseAdapter.commitLog(databaseAdapter.toHash(target))) {
+        assertThat(targetLog).hasSize(i + 1);
+      }
+    }
+
+    // prepare conflict for keys 0 + 1
+
+    Hash conflictBase = databaseAdapter.create(conflict, databaseAdapter.toHash(main));
+    ImmutableCommitAttempt.Builder commit =
+        ImmutableCommitAttempt.builder()
+            .commitToBranch(conflict)
+            .commitMetaSerialized(ByteString.copyFromUtf8("commit conflict"));
+    for (int k = 0; k < 2; k++) {
+      commit.addPuts(
+          KeyWithBytes.of(
+              Key.of("key", Integer.toString(k)),
+              ContentsId.of("C" + k),
+              (byte) 0,
+              ByteString.copyFromUtf8("conflict value for " + k)));
+    }
+    databaseAdapter.commit(commit.build());
+
+    assertThatThrownBy(
+            () -> mergeOrTransplant.apply(conflict, Optional.of(conflictBase), branch, commits, 2))
+        .isInstanceOf(ReferenceConflictException.class)
+        .hasMessage("The following keys have been changed in conflict: 'key.0', 'key.1'");
+
+    return commits;
+  }
+}
